squash! Allow NeuronType to initialize state
This commit is more experimental, but it seems like an improvement.

The main change is to treat the spikes/rates state as the other state
variables, meaning that they're generated in `make_state` and are
passed as **kwargs to the step function.

The reason why it seems like an improvement is because it identified
an edge case with the sig/ops code and the optimizer which would
prevent signals in SimNeurons and SimProcesses from being merged
together properly. I don't believe this happens ever, or is
necessarily possible with the current optimizer, but it seems
plausible that we might want to do that in the future.

I also disabled the progress bar when running tests. While it is
a slight optimization (skips the time printing to stdout), it was
mostly to make it easier to debug builder / optimizer code.
With the progress bar, the `pdb` console is basically unusable
because the progress bar keeps making new lines while trying
to do debug statements.